+++ CHANGELOG.md
@@ -47,8 +47,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Adds oihana\core\strings\dotKebab
 - Adds oihana\core\strings\format
 - Adds oihana\core\strings\formatFromDocument
+- Adds oihana\core\strings\func
 - Adds oihana\core\strings\isRegexp
 - Adds oihana\core\strings\isQuote
+- Adds oihana\core\strings\key
 - Adds oihana\core\strings\pad
 - Adds oihana\core\strings\padBoth
 - Adds oihana\core\strings\padEnd